Remove image link, but leave image
The image link doesn't go to a higher quality version of the image,
so I don't think it adds anything. However, it does mean that
screenreader users or users without a mouse have to click through
an extra option before they can get to the page content.

If folks want to save the image they can right click on it.