Comment out the search box for now, since it's not working.
I'm not convinced a google search like this will ever work for a
one-page site like this. Previously, this search box returned
irelevent results. If we add double-quotes around the "site"
directive so as to truly limit searches to this site, we get zero
results.

It's possible that things might change with time if the google
bots index our PDFs. But for now, I'm removing this search box.

If we want folks to be able to search through all the PDFs and google
still isn't indexing them, we can consider our options later.